projects/research-hub/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "research-hub"
version = "0.1.0"
description = "Academic paper discovery and trend tracking MCP server"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0",
    "supabase>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.hatch.build.targets.wheel]
packages = ["src/research_hub"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
